+++ CHANGELOG.md
@@ -6,6 +6,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased][unreleased]
 
+### Added
+- Initialization with frame.
+- `GMSimplePlayer` with an example that plays sample resource video.
+
 ## [1.1.2](https://github.com/GastonMontes/GMSimplePlayer/releases/tag/1.1.2) - 2017-11-13
 
 ### Added
@@ -26,4 +30,4 @@ All notable changes to this project will be documented in this file.
 ## [1.0.0](https://github.com/GastonMontes/GMSimplePlayer/releases/tag/1.0.0) - 2017-11-04
 
 ### Added
-- `GMSimplePlayer` with and example that plays sample video.
+- `GMSimplePlayer` with an example that plays sample URL video.